pyproject.toml
[project]
name = "kalshi-polymarket-arb"
version = "0.1.0"
description = "Real-time cross-platform arbitrage detection between Kalshi and Polymarket prediction markets"
readme = "README.md"
authors = [
    { name = "ucarberkin", email = "ucarberkin@uchicago.edu" }
]
requires-python = ">=3.11"
dependencies = [
    "websockets>=15.0.0",
    "websocket-client>=1.6.0",
    "cryptography>=41.0.0",
    "matplotlib>=3.5.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

